Create a safe_refresh script
This script is the same as refresh, but first checks the requested url.

This first checks the server to ensure that the page can be loaded successfully.

Suggested use would be periodically in cron, to keep the page up to date. However, if the server goes down, a refresh would result in a blank page, looking bad to end users.

Instead, using this script in cron will ensure that scheduled page refreshes never refresh to a bad page.